git revisions in data dir are working, in a rudimentary way. Still need to prompt for changelog in the UI, and improve the summaries a bit more to indicate the files updated.

+++ addons/revisions/common/src/main/java/org/commonjava/aprox/revisions/RevisionsManager.java
@@ -78,6 +78,11 @@ public class RevisionsManager
     {
         try
         {
+            if ( event.getType() == DataFileEventType.accessed )
+            {
+                return;
+            }
+
             if ( event.getType() == DataFileEventType.deleted )
             {
                 dataFileGit.deleteAndCommit( event.getSummary(), event.getFile() );
